+++ lib/l10n/app_localizations_en.dart
@@ -6331,11 +6331,6 @@ class AppLocalizationsEn extends AppLocalizations {
   }
 
   @override
-  String goalDominantIssueBadge(String dimensionName) {
-    return '$dimensionName needs attention';
-  }
-
-  @override
   String get goalFormAddDimension => 'Add dimension';
 
   @override
@@ -6831,9 +6826,6 @@ class AppLocalizationsEn extends AppLocalizations {
 
   @override
   String get goalReportSectionCoverage => 'Data coverage';
-
-  @override
-  String get goalReportSectionNext => 'What\'s next';
 
   @override
   String get goalReportSectionStanding => 'Where things stand';