+++ lib/l10n/app_localizations_fr.dart
@@ -6463,11 +6463,6 @@ class AppLocalizationsFr extends AppLocalizations {
   }
 
   @override
-  String goalDominantIssueBadge(String dimensionName) {
-    return '$dimensionName a besoin d\'attention';
-  }
-
-  @override
   String get goalFormAddDimension => 'Ajouter une dimension';
 
   @override
@@ -6967,9 +6962,6 @@ class AppLocalizationsFr extends AppLocalizations {
 
   @override
   String get goalReportSectionCoverage => 'Couverture des données';
-
-  @override
-  String get goalReportSectionNext => 'La suite';
 
   @override
   String get goalReportSectionStanding => 'Où tu en es';